pyproject.toml
[project]
name = "obsidian-ops"
version = "0.5.0"
description = "Sandboxed operations on an Obsidian vault."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "Bullish Design", email = "BullishDesignEngineering@gmail.com" },
]
dependencies = [
  "pyyaml>=6.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.5.0",
  "uvicorn>=0.34.0",
]

[project.scripts]
obsidian-ops-server = "obsidian_ops.server:main"

[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/obsidian_ops"]

[tool.hatch.build.targets.sdist]
include = [
  "src",
  "README.md",
  "tests",
  "pyproject.toml",
]

[tool.uv]

[tool.pytest.ini_options]
addopts = "-q --cov=obsidian_ops --cov-report=term-missing"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "UP",
  "B",
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
